requirements.txt
fastapi
uvicorn
openai
gunicorn
pyyaml
python-dotenv